+++ FrontEnd/src/pages/manager/MemberManager.js
@@ -5,10 +5,8 @@ import apiClient from '../../shared/apiClient';
 import { DataGrid } from '@mui/x-data-grid'; // DataGrid 임포트
 import { v4 as uuidv4 } from 'uuid'; // 고유한 ID를 생성하기 위해 uuid 패키지 사용
 import CustomSnackbar from "../../components/common/CustomSnackbar"; // 커스텀 스낵바
-import CustomModal from "../../components/common/CustomModal"; // 커스텀 모달
 
 const MemberManager = () => {
-    const [openModal, setOpenModal] = useState(false);
     const [open, setOpen] = useState(false);
     const [editMode, setEditMode] = useState(false);
     const [selectedEvent, setSelectedEvent] = useState(null);
@@ -243,40 +241,45 @@ const MemberManager = () => {
         }
     };
 
-    const handleDeleteConfirmation = () => {
-        const deletePromises = selectedMembers.map(id => {
-            console.log("Deleting member with ID:", id);
-            return apiClient.delete(`user/${id}`);
-        });
-
-        Promise.all(deletePromises)
-            .then(() => {
-                const updatedEvents = events.filter(event => !selectedMembers.includes(event.id));
-                setEvents(updatedEvents);
-                setSelectedMembers([]);
-                setSnackbarMessage(`${selectedMembers.length}명의 회원 삭제 성공`);
-                setSnackbarSeverity("success");
-                setOpenSnackbar(true);
-            })
-            .catch(error => {
-                console.error('회원 정보를 삭제하지 못했습니다.', error.response.data);
-                setSnackbarMessage('회원 삭제 실패: ' + error.response.data.message);
-                setSnackbarSeverity("error");
-                setOpenSnackbar(true);
-            });
-
-        setOpenModal(false); // 모달 닫기
-    };
-
     // 회원 삭제 핸들러
     const deleteSelectedMembers = () => {
-        if (selectedMembers.length === 0) {
+        // 선택된 회원 수를 확인
+        const memberCount = selectedMembers.length;
+
+        if (memberCount === 0) {
             setSnackbarMessage("삭제할 회원을 선택하세요");
             setSnackbarSeverity("error");
             setOpenSnackbar(true);
             return;
         }
-        setOpenModal(true); // 모달 열기
+
+        const confirmation = window.confirm(`선택된 회원 ${memberCount}명을 삭제하시겠습니까?`); // window.confirm 팝업창
+
+        if (confirmation) {
+            const deletePromises = selectedMembers.map(id => {
+                console.log("Deleting member with ID:", id); // 삭제할 memberId 확인
+                return apiClient.delete(`user/${id}`);
+            });
+
+            Promise.all(deletePromises)
+                .then(() => {
+                    const updatedEvents = events.filter(event => !selectedMembers.includes(event.id));
+                    setEvents(updatedEvents);
+                    setSelectedMembers([]); // 선택한 회원 목록 초기화
+                    setSnackbarMessage(`${memberCount}명의 회원 삭제 성공`);
+                    setSnackbarSeverity("success");
+                    setOpenSnackbar(true);
+                })
+                .catch(error => {
+                    console.error('회원 정보를 삭제하지 못했습니다.', error.response.data);
+                    setSnackbarMessage('회원 삭제 실패: ' + error.response.data.message); // 서버에서 받은 오류 메시지를 출력
+                    setSnackbarSeverity("error");
+                    setOpenSnackbar(true);
+                });
+        } else {
+            // 사용자가 삭제를 취소했을 때의 처리
+            return;
+        }
     };
 
     // 회원 수정 핸들러
@@ -448,65 +451,9 @@ const MemberManager = () => {
                 <Button variant="outlined" onClick={editSelectedMember} sx={{ mr: 2 }}>
                     회원 수정
                 </Button>
-                <div>
-                    <Button variant="outlined" onClick={deleteSelectedMembers} sx={{ mr: 2 }}>
-                        회원 삭제
-                    </Button>
-
-                    {/* Custom Modal */}
-                    <CustomModal isOpen={openModal} closeModal={() => setOpenModal(false)}>
-                        <Box
-                            sx={{
-                                display: "flex",
-                                margin: "auto",
-                                width: "100%",
-                                alignItems: "center",
-                                justifyContent: "center",
-                                flexDirection: "column",
-                                gap: "10px",
-                            }}
-                        >
-                            <h3>회원 삭제</h3>
-                            <p>선택된 회원 {selectedMembers.length}명을 삭제하시겠습니까?</p>
-                            <Box
-                                sx={{
-                                    display: "flex",
-                                    width: "100%",
-                                    alignItems: "center",
-                                    justifyContent: "center",
-                                    flexDirection: "row",
-                                    gap: "24px",
-                                    margin: "16px 0",
-                                }}
-                            >
-                                <Button
-                                    variant="outlined"
-                                    onClick={() => setOpenModal(false)}
-                                    sx={{
-                                        width: "120px",
-                                        height: "40px",
-                                        borderColor: "#34495e",
-                                        color: "#34495e",
-                                    }}
-                                >
-                                    취소
-                                </Button>
-                                <Button
-                                    variant="contained"
-                                    onClick={handleDeleteConfirmation}
-                                    sx={{
-                                        width: "120px",
-                                        height: "40px",
-                                        backgroundColor: "#34495e",
-                                        fontWeight: 600,
-                                    }}
-                                >
-                                    삭제
-                                </Button>
-                            </Box>
-                        </Box>
-                    </CustomModal>
-                </div>
+                <Button variant="outlined" onClick={deleteSelectedMembers}>
+                    회원 삭제
+                </Button>
             </Box>
 
             {/* 모달 */}
